refactor(utils): extract pure remote-URL parsing into git_remote
Split the two side-effect-free remote-URL helpers — slug_from_remote and
web_base_from_remote, plus the three host-matching regexes they own — out
of teatree.utils.git into a new cohesive teatree.utils.git_remote module.

git.py held 37 module-level public functions (over the 10-fn module-health
cap, grandfathered). This is a bounded, behavior-preserving slice: only the
pure string parsers move; the git-invoking operations (run/remote_url/
remote_slug/...) and the GitRepo wrapper stay. The verb functions are NOT
folded onto a class here — that is a separate larger ticket.

git.py: 37 -> 35 public functions, 324 -> 296 LOC (the module-health ratchet
requires only that the over-cap module shrink; it does). The new git_remote
module has 2 public functions, well under the cap.

Every caller updated to the new module (zero stale references, verified by a
repo-wide rg sweep across plain, attribute, and import surface forms):
core/merge/pr_slug_resolution.py, core/reference_linkifier.py,
core/management/commands/_ensure_pr.py, utils/forge.py. The two test classes
move to tests/teatree_utils/test_git_remote.py (the mirror dir for
src/teatree/utils, so the move does not add to the test-path-mirror baseline).

The existing tests pass unchanged against the relocated functions; one case
(unparsable host -> empty web base) was added to cover the final fallthrough.
